Moved sos-, sector- and scenario- model files into folder + tests
