+++ smif/model/composite.py
