+++ smif/model/sector_model.py
@@ -31,10 +31,10 @@ import logging
 import os
 from abc import ABCMeta, abstractmethod
 
-from smif.composite import Model
 from smif.convert.area import get_register as get_region_register
 from smif.convert.interval import get_register as get_interval_register
 from smif.metadata import MetadataSet
+from smif.model.composite import Model
 
 __author__ = "Will Usher, Tom Russell"
 __copyright__ = "Will Usher, Tom Russell"
